% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuyina_underway.R
\name{nuyina_underway}
\alias{nuyina_underway}
\title{Get the Nuyina underway data}
\usage{
nuyina_underway()
}
\value{
data frame
}
\description{
Reads the full Nuyina underway data from a cached URL, and queries any data later than the cache returning all in one
table in datetime order with duplicates removed.
}
\examples{
nuyina_underway()
}
